separate out argparse etc. into cli.py
Also, remove standalone script and use `entry_points` to generate the
executable script.
