--- scripts/beancount-web
@@ -1,15 +1,0 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-
-import os
-import sys
-import argparse
-
-from beancount_web import application
-
-"""
-Run script for beancount-web.
-"""
-
-if __name__ == '__main__':
-    sys.exit(application.run(sys.argv[1:]))
